fixed world snapshots packet deserialization

+++ src/server/src/main.cpp
@@ -19,7 +19,7 @@ int main() {
     Scene world;
     uint32_t next_userID = 100;
     uint32_t next_netID = 1;
-    int frame_counter = 0;
+    uint64_t frame_counter = 0;
 
     uint8_t read_buffer[1024];
     char sender_address[128];
